Refactored GeoLocation function into class and saved coords refactored the Location class, added Geolocation function it saves coordinates to local storage so can be fetched as well as allows coords to be used globally.
