+++ index.html
@@ -6,9 +6,8 @@
     <meta http-equiv="X-UA-Compatible" content="ie=edge">
     <link rel="stylesheet" href="css/styles.css">
     <script src="js/skycons.js"></script>
-    
-    <script src="js/index.js" type="module"></script>
-    <title>Document</title>
+    <script src="js/main.js" type="module"></script>
+    <title>Weather App</title>
 </head>
 <body>
     <div  id="myLocation">
